// gen/api/models/Campaign.ts
/* eslint-disable */


import type { Status } from './Status';

export type Campaign = {
    /**
     * Alphanumeric identifier assigned by the registry for a campaign. This identifier is required by the NetNumber OSR SMS enabling process of 10DLC.
     */
    campaignId?: string;
    /**
     * Alphanumeric identifier of the CSP associated with this campaign.
     */
    cspId?: string;
    /**
     * Alphanumeric identifier of the reseller that you want to associate with this campaign.
     */
    resellerId?: string;
    /**
     * Current campaign status. Possible values: ACTIVE, EXPIRED. A newly created campaign defaults to ACTIVE status.
     */
    status?: Status;
    /**
     * Unix timestamp when campaign was created.
     */
    createDate?: string;
    /**
     * Campaign subscription auto-renewal status.
     */
    autoRenewal?: boolean;
    /**
     * Campaign recent billed date.
     */
    billedDate?: string;
    /**
     * Alphanumeric identifier of the brand associated with this campaign.
     */
    brandId: string;
    /**
     * Business/industry segment of this campaign. Must be of defined valid types. Use `/registry/enum/vertical` operation to retrieve verticals available for given brand, vertical combination.
     */
    vertical: string;
    /**
     * Campaign usecase. Must be of defined valid types. Use `/registry/enum/usecase` operation to retrieve usecases available for given brand.
     */
    usecase: string;
    /**
     * Campaign sub-usecases. Must be of defined valid sub-usecase types. Use `/registry/enum/usecase` operation to retrieve list of valid sub-usecases
     */
    subUsecases?: Array<string>;
    /**
     * Summary description of this campaign.
     */
    description: string;
    /**
     * Does message generated by the campaign include URL link in SMS?
     */
    embeddedLink: boolean;
    /**
     * Does message generated by the campaign include phone number in SMS?
     */
    embeddedPhone: boolean;
    /**
     * Does message content controlled by affiliate marketing other than the brand?
     */
    affiliateMarketing?: boolean;
    /**
     * Does campaign utilize pool of phone nubers?
     */
    numberPool: boolean;
    /**
     * Age gated content in campaign.
     */
    ageGated?: boolean;
    directLending?: boolean;
    /**
     * Does campaign require subscriber to opt-in before SMS is sent to subscriber?
     */
    subscriberOptin: boolean;
    /**
     * Does campaign support subscriber opt-out keyword(s)?
     */
    subscriberOptout: boolean;
    /**
     * Does campaign responds to help keyword(s)?
     */
    subscriberHelp: boolean;
    /**
     * Message sample. Some campaign tiers require 1 or more message samples.
     */
    sample1?: string;
    /**
     * Message sample. Some campaign tiers require 2 or more message samples.
     */
    sample2?: string;
    /**
     * Message sample. Some campaign tiers require 3 or more message samples.
     */
    sample3?: string;
    /**
     * Message sample. Some campaign tiers require 4 or more message samples.
     */
    sample4?: string;
    /**
     * Message sample. Some campaign tiers require 5 or more message samples.
     */
    sample5?: string;
    /**
     * Message flow description.
     */
    messageFlow?: string;
    /**
     * Help message of the campaign.
     */
    helpMessage?: string;
}

export namespace Campaign {

    /**
     * the discriminator for the model if required for more complex api's
     */
    export const discriminator = 'Campaign';


}